+++ examples/dapp-demo/pages/createMsgSend.tsx
@@ -11,8 +11,8 @@ function CreateMsgSend() {
   const [transactionHash, setTransactionHash] = useState<string>()
 
   useEffect(() => {
-    KeplrWallet.init('test').then((wallet) => {
-      GotabitClient.init(wallet, 'test').then(setGotabitInstance)
+    KeplrWallet.init('dev').then((wallet) => {
+      GotabitClient.init(wallet, 'dev').then(setGotabitInstance)
     })
   }, [])
 
@@ -37,10 +37,11 @@ function CreateMsgSend() {
 
     if (!account) throw new Error('Failed to get accounts')
 
+    console.log('to address', toAddress)
     const client = await gotabit.signStargateClient()
     const msg = createMsgGrantBasicAllowance(account, toAddress, [
       {
-        amount: '3000000',
+        amount: '3',
         denom: 'ugtb',
       },
     ])
@@ -60,7 +61,8 @@ function CreateMsgSend() {
     if (!account) throw new Error('Failed to get accounts')
     const client = await gotabit?.signStargateClient()
 
-    const msgSendtoken = createMsgSend(account, toAddress, '3000000', 'ugtb')
+    console.log('to address', toAddress, account)
+    const msgSendtoken = createMsgSend(account, toAddress, '3', 'ugtb')
     const result = await client.signAndBroadcast(
       account,
       [msgSendtoken],