+++ examples/dapp-demo/pages/wallet-gotabit-pc.tsx
@@ -1,24 +1,24 @@
-import { useState, ChangeEvent } from 'react'
+import { useState, useEffect, ChangeEvent } from 'react'
 import { GotabitWallet } from '@gotabit/wallet-gotabit'
-import { createMsgSend } from '@gotabit/messages'
 import { GotabitClient } from '@gotabit/client'
+import { createMsgSend } from '@gotabit/messages'
 
-function WalletconnectPage() {
-  const [account, setAccount] = useState<string>()
-
+function WalletGotabit() {
   const [toAddress, setToAddress] = useState<string>('')
   const [responseData, setResponseData] = useState<any>()
   const [gotabit, setGotabitInstance] = useState<GotabitClient>()
   const [transactionHash, setTransactionHash] = useState<string>()
+  const [account, setAccount] = useState<string>()
 
+  useEffect(() => {
+    GotabitWallet.init('dev').then((wallet) => {
+      GotabitClient.init(wallet, 'dev').then(setGotabitInstance)
+    })
+  }, [])
   const handleConnect = async () => {
-    const wallet = await GotabitWallet.init('test')
-
-    GotabitClient.init(wallet, 'test').then(setGotabitInstance)
+    const wallet = await GotabitWallet.init('dev')
 
     const accounts = await wallet.getAccounts()
-
-    console.log('---accounts', accounts);
 
     if (accounts?.[0]) setAccount(accounts?.[0].address)
   }
@@ -32,7 +32,7 @@ function WalletconnectPage() {
     if (!account) throw new Error('Failed to get accounts')
     const client = await gotabit?.signStargateClient()
 
-    const msgSendtoken = createMsgSend(account, toAddress, '30', 'ugtb')
+    const msgSendtoken = createMsgSend(account, toAddress, '3', 'ugtb')
     const result = await client.signAndBroadcast(
       account,
       [msgSendtoken],
@@ -49,15 +49,16 @@ function WalletconnectPage() {
 
   return (
     <div>
-      <button onClick={handleConnect}>connect</button>
-      <p>Address: {account}</p>
-      <br />
-      <div>
+      <p>
         <label>To Address</label>
         <input type="text" onChange={handleChange} />
-      </div>
+      </p>
+      <br />
+      <button onClick={handleConnect}>connect</button>
+      <p>Address: {account}</p>
       <button onClick={sendTokenWithMsgSend}>test msgSend</button>
-      <div>
+      <br />
+      <p>
         <a
           href={`https://explorer.hjcore.io/GotaBit-test/tx/${transactionHash}`}
           target="_blank"
@@ -65,10 +66,10 @@ function WalletconnectPage() {
         >
           See your transation in explorer
         </a>
-      </div>
+      </p>
       <p>result: {JSON.stringify(responseData)}</p>
     </div>
   )
 }
 
-export default WalletconnectPage
+export default WalletGotabit