+++ packages/messages/src/txRaw.ts
@@ -1,22 +1,20 @@
-import { cosmos } from '@gotabit/proto'
-
-const tx = cosmos.tx.v1beta1
+import { TxRaw } from 'cosmjs-types/cosmos/tx/v1beta1/tx'
 
 export function createTxRaw(
   bodyBytes: Uint8Array,
   authInfoBytes: Uint8Array,
   signatures: Uint8Array[],
 ) {
-  const message = tx.TxRaw.fromPartial({
+  const message = TxRaw.fromPartial({
     bodyBytes,
     authInfoBytes,
     signatures,
   })
 
-  const msgBytes = Uint8Array.from(tx.TxRaw.encode(message).finish())
+  const msgBytes = Uint8Array.from(TxRaw.encode(message).finish())
 
   return {
-    message: msgBytes,
+    value: msgBytes,
     typeUrl: '/cosmos.tx.v1beta1.TxRaw',
   }
 }