+++ packages/wallet-core/src/utils.ts
@@ -1,7 +1,6 @@
-import { GasPrice } from '@cosmjs/stargate';
 import * as secp from '@noble/secp256k1';
 import { pubkeyToAddress, serializeSignDoc, StdSignDoc } from '@cosmjs/amino';
-
+import { Decimal } from '@cosmjs/math';
 import { stringToPath, sha256 } from '@cosmjs/crypto';
 import {
   fromBase64,
@@ -11,8 +10,16 @@ import {
   fromBech32,
 } from '@cosmjs/encoding';
 import { makeSignBytes } from '@cosmjs/proto-signing';
-import { SignDoc } from 'cosmjs-types/cosmos/tx/v1beta1/tx';
 import Long from 'long';
+
+import {
+  SignDoc,
+  ConfigType,
+  GotaBitConfig,
+  ConfigTypeEnum,
+  GotaBitWalletOptions,
+  GotaBitInitWalletOptions,
+} from './types';
 
 import {
   TEST_CONFIG,
@@ -23,13 +30,62 @@ import {
   DEFAULT_ADDRESS_PREFIX,
   DEV_CONFIG,
 } from './constants';
-import {
-  ConfigType,
-  GotaBitConfig,
-  ConfigTypeEnum,
-  GotaBitWalletOptions,
-  GotaBitInitWalletOptions,
-} from './types';
+
+/**
+ * Denom checker for the Cosmos SDK 0.42 denom pattern
+ * (https://github.com/cosmos/cosmos-sdk/blob/v0.42.4/types/coin.go#L599-L601).
+ *
+ * This is like a regexp but with helpful error messages.
+ */
+function checkDenom(denom: string): void {
+  if (denom.length < 3 || denom.length > 128) {
+    throw new Error('Denom must be between 3 and 128 characters');
+  }
+}
+
+/**
+ * A gas price, i.e. the price of a single unit of gas. This is typically a fraction of
+ * the smallest fee token unit, such as 0.012utoken.
+ */
+export class GasPrice {
+  public readonly amount: Decimal;
+
+  public readonly denom: string;
+
+  public constructor(amount: Decimal, denom: string) {
+    this.amount = amount;
+    this.denom = denom;
+  }
+
+  /**
+   * Parses a gas price formatted as `<amount><denom>`, e.g. `GasPrice.fromString("0.012utoken")`.
+   *
+   * The denom must match the Cosmos SDK 0.42 pattern (https://github.com/cosmos/cosmos-sdk/blob/v0.42.4/types/coin.go#L599-L601).
+   * See `GasPrice` in @cosmjs/stargate for a more generic matcher.
+   *
+   * Separators are not yet supported.
+   */
+  public static fromString(gasPrice: string): GasPrice {
+    // Use Decimal.fromUserInput and checkDenom for detailed checks and helpful error messages
+    const matchResult = gasPrice.match(/^([0-9.]+)([a-z][a-z0-9]*)$/i);
+    if (!matchResult) {
+      throw new Error('Invalid gas price string');
+    }
+    const [_, amount, denom] = matchResult;
+    checkDenom(denom);
+    const fractionalDigits = 18;
+    const decimalAmount = Decimal.fromUserInput(amount, fractionalDigits);
+    return new GasPrice(decimalAmount, denom);
+  }
+
+  /**
+   * Returns a string representation of this gas price, e.g. "0.025uatom".
+   * This can be used as an input to `GasPrice.fromString`.
+   */
+  public toString(): string {
+    return this.amount.toString() + this.denom;
+  }
+}
 
 export interface ChainConfig extends GotaBitConfig {
   chainType: ConfigType;