+++ packages/qrcode-modal/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.18](https://github.com/hjcore/sdk-ts/compare/@gotabit/qrcode-modal@1.1.17...@gotabit/qrcode-modal@1.1.18) (2022-11-21)
+
+**Note:** Version bump only for package @gotabit/qrcode-modal
+
+## [1.1.17](https://github.com/hjcore/sdk-ts/compare/@gotabit/qrcode-modal@1.1.16...@gotabit/qrcode-modal@1.1.17) (2022-11-19)
+
+**Note:** Version bump only for package @gotabit/qrcode-modal
+
+## [1.1.16](https://github.com/hjcore/sdk-ts/compare/@gotabit/qrcode-modal@1.1.15...@gotabit/qrcode-modal@1.1.16) (2022-11-11)
+
+**Note:** Version bump only for package @gotabit/qrcode-modal
+
+## [1.1.15](https://github.com/hjcore/sdk-ts/compare/@gotabit/qrcode-modal@1.1.14...@gotabit/qrcode-modal@1.1.15) (2022-11-06)
+
+**Note:** Version bump only for package @gotabit/qrcode-modal
+
+## [1.1.14](https://github.com/hjcore/sdk-ts/compare/@gotabit/qrcode-modal@1.1.12...@gotabit/qrcode-modal@1.1.14) (2022-11-05)
+
+**Note:** Version bump only for package @gotabit/qrcode-modal
+
 ## [1.1.12](https://github.com/hjcore/sdk-ts/compare/@gotabit/qrcode-modal@1.1.11...@gotabit/qrcode-modal@1.1.12) (2022-11-04)
 
 **Note:** Version bump only for package @gotabit/qrcode-modal