+++ packages/wallet-walletconnect/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.20](https://github.com/gotabit/sdk-ts/compare/@gotabit/wallet-walletconnect@1.1.19...@gotabit/wallet-walletconnect@1.1.20) (2022-11-21)
+
+**Note:** Version bump only for package @gotabit/wallet-walletconnect
+
+## [1.1.19](https://github.com/gotabit/sdk-ts/compare/@gotabit/wallet-walletconnect@1.1.18...@gotabit/wallet-walletconnect@1.1.19) (2022-11-19)
+
+**Note:** Version bump only for package @gotabit/wallet-walletconnect
+
+## [1.1.18](https://github.com/gotabit/sdk-ts/compare/@gotabit/wallet-walletconnect@1.1.17...@gotabit/wallet-walletconnect@1.1.18) (2022-11-11)
+
+**Note:** Version bump only for package @gotabit/wallet-walletconnect
+
+## [1.1.17](https://github.com/gotabit/sdk-ts/compare/@gotabit/wallet-walletconnect@1.1.16...@gotabit/wallet-walletconnect@1.1.17) (2022-11-10)
+
+**Note:** Version bump only for package @gotabit/wallet-walletconnect
+
+## [1.1.16](https://github.com/gotabit/sdk-ts/compare/@gotabit/wallet-walletconnect@1.1.15...@gotabit/wallet-walletconnect@1.1.16) (2022-11-10)
+
+**Note:** Version bump only for package @gotabit/wallet-walletconnect
+
+## [1.1.15](https://github.com/gotabit/sdk-ts/compare/@gotabit/wallet-walletconnect@1.1.14...@gotabit/wallet-walletconnect@1.1.15) (2022-11-06)
+
+**Note:** Version bump only for package @gotabit/wallet-walletconnect
+
+## [1.1.14](https://github.com/gotabit/sdk-ts/compare/@gotabit/wallet-walletconnect@1.1.12...@gotabit/wallet-walletconnect@1.1.14) (2022-11-05)
+
+**Note:** Version bump only for package @gotabit/wallet-walletconnect
+
 ## [1.1.12](https://github.com/gotabit/sdk-ts/compare/@gotabit/wallet-walletconnect@1.1.11...@gotabit/wallet-walletconnect@1.1.12) (2022-11-04)
 
 **Note:** Version bump only for package @gotabit/wallet-walletconnect